+++ ELOR.VKAPILib/Objects/Album.cs
@@ -1,9 +1,4 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ELOR.VKAPILib.Objects {
     public class Album {
@@ -11,7 +6,7 @@ namespace ELOR.VKAPILib.Objects {
         public int Id { get; set; }
 
         [JsonProperty("owner_id")]
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
 
         [JsonProperty("title")]
         public string Title { get; set; }
