+++ ELOR.VKAPILib/Objects/Attachment.cs
@@ -1,10 +1,5 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.Serialization;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ELOR.VKAPILib.Objects {
     [DataContract]
@@ -86,7 +81,7 @@ namespace ELOR.VKAPILib.Objects {
         public virtual string ObjectType { get; set; }
 
         [JsonProperty("owner_id")]
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
 
         [JsonProperty("id")]
         public int Id { get; set; }
