+++ ELOR.VKAPILib/Objects/Call.cs
@@ -1,15 +1,10 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ELOR.VKAPILib.Objects
 {
     public class CallParticipants {
         [JsonProperty("list")]
-        public List<int> List { get; set; }
+        public List<long> List { get; set; }
 
         [JsonProperty("count")]
         public int Count { get; set; }
@@ -17,10 +12,10 @@ namespace ELOR.VKAPILib.Objects
 
     public class Call {
         [JsonProperty("initiator_id")]
-        public int InitiatorId { get; set; }
+        public long InitiatorId { get; set; }
 
         [JsonProperty("receiver_id")]
-        public int ReceiverId { get; set; }
+        public long ReceiverId { get; set; }
 
         [JsonProperty("state")]
         public string State { get; set; }
@@ -40,7 +35,7 @@ namespace ELOR.VKAPILib.Objects
 
     public class GroupCallInProgress {
         [JsonProperty("initiator_id")]
-        public int InitiatorId { get; set; }
+        public long InitiatorId { get; set; }
 
         [JsonProperty("participants")]
         public CallParticipants Participants { get; set; }
