+++ ELOR.VKAPILib/Objects/ChatPreview.cs
@@ -1,21 +1,16 @@
 ﻿
 using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ELOR.VKAPILib.Objects {
     public class ChatPreview {
         [JsonProperty("admin_id")]
-        public int AdminId { get; set; }
+        public long AdminId { get; set; }
 
         [JsonProperty("members_count")]
         public int MembersCount { get; set; }
 
         [JsonProperty("members")]
-        public List<int> Members { get; set; }
+        public List<long> Members { get; set; }
 
         [JsonProperty("title")]
         public string Title { get; set; }
@@ -24,7 +19,7 @@ namespace ELOR.VKAPILib.Objects {
         public Photo Photo { get; set; }
 
         [JsonProperty("local_id")]
-        public int LocalId { get; set; }
+        public long LocalId { get; set; }
 
         [JsonProperty("joined")]
         public bool Joined { get; set; }
