+++ ELOR.VKAPILib/Objects/PhotoSaveResult.cs
@@ -1,9 +1,4 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ELOR.VKAPILib.Objects {
     public class PhotoSaveResult {
@@ -17,7 +12,7 @@ namespace ELOR.VKAPILib.Objects {
         public int AlbumId { get; set; }
 
         [JsonProperty("owner_id")]
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
 
         [JsonProperty("access_key")]
         public string AccessKey { get; set; }
