+++ L2/Converters/AvatarGradientConverter.cs
@@ -8,7 +8,7 @@ using System.Globalization;
 namespace ELOR.Laney.Converters {
     public sealed class AvatarGradientConverter : IValueConverter {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture) {
-            if (value is int id) {
+            if (value is long id) {
                 return id.GetGradient();
             }
             return new SolidColorBrush(Colors.Transparent);
